- added presenter, which handles immediate mode rendering, ui, input etc. anything related to interface with players and UX - fixed all include since the folder changes - increased warning level and made warning as errors, fixed all issues arose

+++ rogue/source/engine/shader.cpp
@@ -121,7 +121,7 @@ Shader::~Shader()
 {
 	if (shader_id > 0)
 	{
-		glDeleteProgram(shader_id);
+		glDeleteProgram(static_cast<GLuint>(shader_id));
 	}
 }
